refactor: separating default values ​​from the form
