+++ frontend/src/pages/(protected)/pacientes/_components/actions/add-patient-button.tsx
@@ -3,7 +3,7 @@ import { useState } from 'react';
 
 import { Button } from '@/components/ui/button';
 import { Dialog, DialogTrigger } from '@/components/ui/dialog';
-import { UpsertPatientForm } from './upsert-patient-form';
+import { UpsertPatientForm } from '../upsert-patient-form';
 
 export const AddPatientButton = () => {
   const [isOpen, setIsOpen] = useState(false);
